requirements.txt
streamlit
pdfplumber
google-genai
python-dotenv
markdown
xhtml2pdf
